# tests/api/openapi_v1_snapshot.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "The consistent error envelope returned for handled failures (AAEO-9).",
        "properties": {
          "code": {
            "default": "error",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EvaluationRun": {
        "description": "One execution of a suite against a set of actual agent outputs.\n\n``metadata`` carries derived diagnostics (e.g. ``calibration_ece``, AAEO-5)\nand a ``persisted`` flag when storage write fails (AAEO-15). ``workspace_id``\ncarries the tenant the run was scored under so reads can be tenant-scoped\n(AAEO-4) \u2014 ``None`` on the offline/single-tenant path and for legacy rows\nwritten before this field existed.",
        "properties": {
          "aggregate_score": {
            "default": 0.0,
            "title": "Aggregate Score",
            "type": "number"
          },
          "case_results": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvaluationCaseResult"
            },
            "title": "Case Results",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "suite_id": {
            "title": "Suite Id",
            "type": "string"
          },
          "suite_name": {
            "title": "Suite Name",
            "type": "string"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "suite_id",
          "suite_name"
        ],
        "title": "EvaluationRun",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "x-framework-generated": "validation-error-envelope"
      },
...
      "RunSuiteRequest": {
        "description": "The POST /v1/evaluation/runs body: a suite + per-case actual outputs.\n\n``workspace_id`` scopes the run to its owning tenant (resolved against the\nprincipal, AAEO-4); it is optional and ``None`` on the offline/single-tenant\npath, preserving legacy behavior.",
        "properties": {
          "actual_outputs": {
            "additionalProperties": true,
            "title": "Actual Outputs",
            "type": "object"
          },
          "suite": {
            "$ref": "#/components/schemas/EvaluationSuite"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "suite"
        ],
        "title": "RunSuiteRequest",
        "type": "object"
      },
...
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/evaluation/runs": {
      "get": {
        "description": "List evaluation runs, optionally filtered by suite id (tenant-scoped, AAEO-4).",
        "operationId": "list_evaluation_runs_v1_evaluation_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "suite_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Suite Id"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationRun"
                  },
                  "title": "Response List Evaluation Runs V1 Evaluation Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Evaluation Runs",
        "tags": [
          "evaluation"
        ]
      },
      "post": {
        "description": "Score a suite against actual outputs and persist + return the run.\n\nThe run is stamped with the resolved ``workspace_id`` (AAEO-4) and the creating\nactor (audit spine, T1.1) so later reads can be tenant-scoped.",
        "operationId": "run_suite_v1_evaluation_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSuiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Suite",
        "tags": [
          "evaluation"
        ]
      }
    },
    "/v1/evaluation/runs/{run_id}": {
      "get": {
        "description": "Read one evaluation run by id (404 when unknown/out-of-workspace, AAEO-4).",
        "operationId": "get_evaluation_run_v1_evaluation_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evaluation Run",
        "tags": [
          "evaluation"
        ]
      }
    },
...
    "/v1/runs": {
      "get": {
        "description": "List runs (created_at desc), paginated. Returns a paged envelope (AAEO-8).\n\nObject-level (BOLA, WS-bola): a ``subject_scoped`` principal's list is pinned to its\nOWN subject via :func:`narrow_subject` (a ``tenant_admin`` and every non-subject-scoped\n/ offline caller keep the caller-supplied filter), so a subject-scoped tenant can never\nenumerate another data subject's runs within a shared workspace.",
        "operationId": "list_runs_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "in": "query",
            "name": "subject_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Create a run (idempotent) and execute it in the background; returns the record.\n\nTwo mutually-exclusive identity sources (T2e): an inline ``persona`` (tool-less,\nbyte-unchanged back-compat) OR an ``agent_id`` that resolves a stored ``AgentSpec``\nand executes WITH ITS TOOLS on the per-run runtime, recording an agent lineage edge.\n\nWith ``hitl=true`` (T2f) the run drives the agentic loop and PAUSES at\n``AWAITING_APPROVAL`` when it hits an approval-gated tool \u2014 resumable via\n``POST /v1/runs/{id}/approve|reject``. HITL requires a stored ``agent_id`` (an inline\npersona carries no tools to gate \u2192 422) and a deployment with a checkpoint store\n(\u2192 400 when absent).\n\nWith ``plan=true`` (T2g) the run pauses at PLAN-READY: the agent publishes its plan\n(carried in ``metadata['plan']``) and the run stops at ``AWAITING_APPROVAL``,\nresumable via the SAME ``approve`` endpoint. Plan mode has the same preconditions as\nHITL (a stored ``agent_id`` + a checkpoint store).",
        "operationId": "create_run_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "description": "Read one run record by id (404 when unknown/out-of-workspace, AAEO-4).\n\nObject-level (BOLA, WS-bola) narrowing rides on top of the workspace scope: a\n``subject_scoped`` principal that resolves a run belonging to ANOTHER data subject\nwithin its own tenant gets a clean 404 (not-found, never a 403 that would confirm the\nrun exists). Non-subject-scoped / ``all_tenants`` / offline callers are unaffected \u2014\n:func:`authorize_object` is a no-op for them, so a shared multi-user workspace and the\nzero-config path keep reading every run byte-for-byte.",
        "operationId": "get_run_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}